requirements.txt
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.3
decorator==5.1.1
Django==5.0.4
django-debug-toolbar==4.3.0
django-extensions==3.2.3
executing==2.0.1
feedparser==6.0.11
ipython==8.23.0
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg==3.1.18
ptyprocess==0.7.0
pure-eval==0.2.2
pydotplus==2.0.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
typing_extensions==4.11.0
wcwidth==0.2.13
